backend/tests/conftest.py
"""Pytest bootstrap: load .env so tests pick up LLM credentials when present."""
from __future__ import annotations

from pathlib import Path

from dotenv import load_dotenv

_REPO_ROOT = Path(__file__).resolve().parent.parent
load_dotenv(_REPO_ROOT / ".env", override=False)
