backend/pyproject.toml
[project]
name = "kairo-backend"
version = "0.1.0"
description = "FastAPI backend for Kairo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.135.3",
  "httpx==0.28.1",
  "langchain-anthropic==1.4.0",
  "langchain-ollama==1.1.0",
  "langchain-openai==1.1.12",
  "langgraph==1.1.6",
  "langgraph-checkpoint-sqlite>=3.0.3",
  "aiosqlite==0.20.0",
  "pydantic==2.13.0",
  "python-dotenv==1.2.2",
  "uvicorn[standard]==0.44.0",
  "google-api-python-client>=2.194.0",
  "google-auth>=2.49.2",
  "google-auth-oauthlib>=1.3.1",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3",
  "ruff==0.11.13",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["assistant"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
